refactor: organize change request components
