+++ frontend/src/components/ChangeRequests/BudgetChangeReviewCard/BudgetChangeReviewCard.jsx
@@ -1,8 +1,8 @@
 import PropTypes from "prop-types";
 import { convertCodeForDisplay, renderField } from "../../../helpers/utils";
 import { useGetNameForCanId, useGetBLIStatus } from "../../../hooks/lookup.hooks";
-import ReviewCard from "./ReviewCard";
-import TermTag from "./TermTag";
+import ReviewCard from "../ReviewCard/ReviewCard";
+import TermTag from "../TermTag";
 /**
  * BudgetChangeReviewCard component
  * @component
