+++ frontend/src/components/ChangeRequests/BudgetChangeReviewCard/index.js
@@ -1,0 +1,2 @@
+export * from "./BudgetChangeReviewCard";
+export { default } from "./BudgetChangeReviewCard";
