+++ frontend/src/components/ChangeRequests/StatusChangeReviewCard/index.js
@@ -1,0 +1,2 @@
+export * from "./StatusChangeReviewCard";
+export { default } from "./StatusChangeReviewCard";
