+++ frontend/src/components/ChangeRequests/TermTag/index.js
@@ -1,0 +1,2 @@
+export * from "./TermTag";
+export { default } from "./TermTag";
